+++ org.yamcs.studio.core/src/main/java/org/yamcs/studio/data/vtype/ListDouble.java
@@ -63,27 +63,27 @@ public abstract class ListDouble implements ListNumber, CollectionDouble {
 
     @Override
     public void setFloat(int index, float value) {
-        setDouble(index, (double) value);
+        setDouble(index, value);
     }
 
     @Override
     public void setLong(int index, long value) {
-        setDouble(index, (double) value);
+        setDouble(index, value);
     }
 
     @Override
     public void setInt(int index, int value) {
-        setDouble(index, (double) value);
+        setDouble(index, value);
     }
 
     @Override
     public void setShort(int index, short value) {
-        setDouble(index, (double) value);
+        setDouble(index, value);
     }
 
     @Override
     public void setByte(int index, byte value) {
-        setDouble(index, (double) value);
+        setDouble(index, value);
     }
 
     /**
